+++ .github/workflows/openmldb-tool.yml
@@ -39,6 +39,7 @@ jobs:
           bash test/test-tool/openmldb-deploy/install.sh ${VERSION}
       - name: run test
         run: |
+          cp -f tools/tool.py test/test-tool/openmldb-deploy/cases/
           python3 -m pytest test/test-tool/openmldb-deploy/cases --junit-xml=pytest.xml
       - name: clear env
         run: |
@@ -71,6 +72,7 @@ jobs:
           bash test/test-tool/openmldb-deploy/install_with_name.sh ${VERSION}
       - name: run test
         run: |
+          cp -f tools/tool.py test/test-tool/openmldb-deploy/cases/
           python3 -m pytest test/test-tool/openmldb-deploy/cases --junit-xml=pytest.xml
       - name: clear env
         run: |
